add option to set run_type in raw_data_manip and switch to docopt
Adds the --run_type option, which enables the user to give a string
describing the run type (Calib or Data), which will be added to the
name of the output file. Thus, also change the dump output proc.

Switches the command line argument parsing to docopt, which given a CL
docstring, will create a table of command line arguments.