introduce batchFiles template
Introduces the batchFiles template, which can be used to arbitrarily
batch work on a list of files. Hand it a list of files and you can
access any number (given to the template) of files at a time,
e.g. provide a read function to it and then work on the read data.