add proc get tracking events + filter tracking indices for chips
Adds a proc to get the event numbers corresponding to tracking or non
tracking.
Also adds a proc to filter out the correct indices for different
chips, which obviously don't all have clusters in each event, i.e. get
a mapping of event numbers to indices for each chips' datasets.