batch InGrid files using `batchFiles`
In order to accomodate that, the FADC and InGrid procs were properly
split. Instead of receiving a name of a `run_folder`,
`readRawInGridData` now receives a `listOfFiles` which it works on.

We read all files in the run folder sorted by filename, batch these
files, then sort the batches by inode.

The writing was completely restructured. Split into an `init` phase,
a `write attributes` phase and a `write hyperslab` phase and finally a
`create hyperlinks` phase.

To be able to write hyperslabs of variable length data `nimhdf5` had
to be updated.