open / close H5 file after each run / currently call subprocess
Due to a memory leak somewhere in `raw_data_manipulation` we simply
call a subprocess for each sub folder, which is a TOS run folder
instead of calling the proc to process a single run.

Still, in the commented code, the opening and closing of the H5 file
was moved before / after the individual processings of each run to
clean up.