[Untested] change the database read and write to float for SCurves
In accordance with the previous commit also change the data written
and read to/from the database for SCurves to also be floats.
NOTE: this a breaking change with the previous database. We won't be
able to read from the old database anymore, since it will crash when
trying to read an int dataset as a float.

We could introduce something similar to what is done in arraymancer's
HDF5 interface, where we make use of nimhdf5's `convertType` proc.