[LH] add Christoph PhD thesis LH cut values, ref #18
Adds a proc for the exact values Christoph used in this PhD thesis, so
that we can exclude the possibility of more bugs related to
determination of the actual cut values.

The resulting plot shows that the difference is really miniscule (at
least not actually related to the bug we still have).