[raw] fix usage of raw for folder of runs, if outfile non existing
With the previous version of the raw data manipulation using it on a
folder of run folders on an output file that isn't existing, the
program would fail with the message that a non existing file cannot be
created if opened with `"r"`. Now we check whether the file exists,
and if it does not, we still open with "rw".