[calib] add `cutOnDsets` helper proc to simplify cutting on datasets
Adds a general proc to cut on datasets, which one can give a variable
number of arguments, besides some required data (event numbers, chip
region to cut to, centerX, centerY and number of hits). For each
additional dataset a lower and upper bound is given, which is then
checked for.