[raw] add `getRunHeader` to return run header based on Events
Previously we extracted the run header simply from the 0th element of
the `seq[Event]` within the `processRawInGridData` proc, which was
completely arbitrary. Now at least we do it at one clear point in the
code, although it is still hardcoded.

Done for extension for SRS TOS later, where we cannot do it in the old
way, since shutter information is not available event wise.