[reco] fix FADC property calculation, not dropping last event
During the calculation of the FADC event properties during the
reconstruction phase (using --only_fadc), previously we dropped (in
some cases?) the very last element of the FADC data due to a bad
reshaping of the FADC data (which is returned as a 1D array from
nimhdf5, despite being a 2D array).

Fixes this by removing the badly implemented `reshapeFADC` proc.

Also gets rid of the not used `minvals` dataset in the FADC
calculation proc.