major WIP update staticClient parse string to JsObject instead...
of parsing the JSON string to a Nim JsonNode, which we later have to
convert back to a string before parsing it to a JsObject *again*.

This *significantly* improves the start up time of the program as well
as the time it takes to plot an individual plot!

Plotly plots are now included in a 'foreign node', which are created
after the rendering of karax is done.