[python] use `binning` as x elements for FeSpectrum fit functions
No point in using linspace, if we only intend to use about as many as
we have bins anyways. Yes, the fit will technically only correspond to
the bin edges, but that doesn't really matter, since the function is
used to calculate the values based on the best fit and the points will
also be plotted at the bin edges and not the centers, as before.