[plot] add prototype for string evaluation of functions
With this prototype we can read a string from the karax frontend, send
it to the server, evaluate it offline and send the results as a plot
back to the front end.

Alternatively one could of course perform certain plots from JS
directly, by giving a user interface to the plotly.