[reco] fix Fe spectrum calls for SRS w/o real chip name
If a calibration run was handled by the reconstruction for the SRS TOS
without a `run.txt` / no Chip ID, we previously raised a `ValueError`,
due to a name mismatch when checking for the existence of the chip in
the InGrid database.

Also makes the "SRS Chip" name a constant for which we check in `inDatabase`.