[plot] run websocket server in 2nd thread, send messages chunked
The websocket server now runs in a separate thread, which allows for
easier handling of multiple clients (although at the moment on the
last is being served, I believe).

Secondly, for some reason Chromium does not like websocket frames,
which are larger than 32kB.
Thus, we split each data packet into smaller fragments and concat the
data again on the receiving side.