[plotting] add script to create OuterChips hist plot
Serves as an example how to use `karaPlot` as a library to create
plots.

The script creates the histogram for the number of hits of events on
the outer chips, in case there was an event on the center chip which
is "blob" like, meaning passing simple cuts on eccentricity and
rmsTransverse.

Background and calibration H5 files are combined and the histograms overlaid.