add WIP code for an online event display + hits histo
The code so far watches a run folder, reads the files in it already,
reconstructs them, writes them to a H5 file and reads + recos files
being written to the directory and adds the reconstructed to the H5
file too.

Remaining is creation of the plot (and redrawing diffs) and writing a
simple client running on JS backend to show the plotting.

NOTE: this currently depends on some non commited code in the repo +
the `fsmonitor2` package.