[cdl] add ability to fix variables based on refs to other params
Allows to fix a fit parameter by referencing other parameters. For
instance considering a fit consisting of two gaussians, we can
reference other paramters (potentially from other lines) to define
some calculation to do with those:
```nim
declareFitFunc(cuNi15):
  ffGauss: "Cu-Kalpha"
  ffGauss:
    name = "Cu-esc"
    gmu = gmu("Cu-Kalpha") * gs("Cu-Kalpha")
```
In this case the parameter `gmu` of the `Cu-esc` line will be
fixed. But not to a fixed value, but rather to the product of two
other fit parameters. The resulting fit function will then look like:
```nim
proc cuNi15Func(p_ar: seq[float], x: float): float =
  result += p_ar[0] * exp(x, p_ar[1], p_ar[2])
  result += p_ar[3] * exp(x, p_ar[1] * p_ar[2], p_ar[4])
```

Note that the parameter still has to be set as `fixed` in the
`getLines` procs!