[cdl] finish `generateCdlCalibrationFile` implementation
In principle the implementation should work. However, I don't have the
correct H5 input available with me right now, so I can't test that
it's actually quite right.

There's one caveat at the moment:
When creating the calibration-cdl.h5 file for the 2014 data from the
raw data reconstructed with TimepixAnalsis, the names of the datasets
in the groups will be the same as for the later data, since we just
reuse the names from the input.
This will break the usage of genRefFile based on the thus created H5
calibration-cdl, since that (with the "apr2014" string contained) will
expect the old dataset names. The fix is simple, but will have to wait
for now.