[cdl] make `genRefFile` from 2018 calibration via `genCdlFile` work
Only slight changes were necesary.

NOTE: this is still missing one crucial point: that is the charge cut
values we currently apply to the data are the same ones we use for the
2014 data. Since our detector obviously ran at different gas gains
etc, those charge values do not map 1:1. Need to look at spectra and
then modify the proc to accomodate for that.