fix embarrassing totalCharge bug
Calculation of total charge (as I suspected in a recent commit, in
which I added a `doAssert`) was wrong. Since the calibration function
is a non-linear function, it makes no sense to use the sum of the ToT
valus for the calculation of the total charge. Instead it must be the
sum of the individual charges!

I changed this months ago on my laptop locally, but never commited
it. Damn...!