[reco] reorder usage flags in order they should be run
The flags are now roughly shown in the order in which they should be
run by the user for a correct full analysis. Roughly, because of two
things:
- create_fe_spec is only applicable to calibration data and is done
automatically for those. The flag is only useful if a run is not
stored as a calibration run (has no `runType`) or if it's to be run on
any other run type
- `only_energy <factor>` is redundant / an alternative to "only_energy_from_e`