[calib] MAJOR, fixes application of filtered Idxs to pix <-> events
This was a major bug in the calculation of the gas gain. Instead of
applying the filtered event indices to the events and then taking the
pixels from those events passing the cuts, we errorneously flattened
the read ToT data from seq[seq[uint16]] to seq[uint16] and applied the
filtering to the resulting 1D sequence. This had the effect that
instead of filtering out individual events, we filtered to *individual
pixels*! This obviously reduced our statistics by 2-3 orders of
magnitude, resulting in significantly worse predicitons of gas gain.

This is explained in more detail in the `statusAndProgress.org` file!