[calib] use the mean of data histogram as gas gain instead of fit
The mean of the fit will be kept as `G_fitmean`. The default `G` is
now the mean of the data histogram. This is in line with what happens
in Krieger PhD thesis.

The result of this is a much better agreement between the histograms
of the gas gains comparing Marlin 2014 w/ TPA 2014.

See `statusAndProgress.org` for details.