MAJOR: fix major bug concerning total back. time in SPSC_Sep_2020 tag
This is an amendment to a previous commit. This will be force pushed
to avoid confusing my future self with a wrong statement in the
future.

The commit message was the same, but my understanding was wrong while
fixing the code below. This was the old fix:

```nim
     let lhGrp = h5f["/likelihood".grp_str]
     result = lhGrp.attrs["totalDuration", float]

-  let trackingTime = h5Cands.mapIt(it.readDuration).sum / 10.0
+  let trackingTime = h5Cands.mapIt(it.readDuration).sum
   echo "Total tracking time ", trackingTime / 3600.0, " h"
   let secondsOfSim = N_sim.float / totalFluxPerYear * 86400 * 365
...
  let scale = trackingTime.float / secondsOfSim / (100 * 100) * areaBore
  echo &"Scale = {scale}"
  let gaeDf = readAxModel(axionModel, scale)
```

The explanation as to why this was a bad fix:

*WARNING2*: Turns out past me once again was not as stupid as the not
as past me thought. Instead of removing that factor of 10 outright,
I have to modify it of course. That factor is the
 *ratio of tracking to non tracking time*! My mind yesterday read
=trackingTime= as =backgroundTime=. Of course we need to scale the the
signal hypothesis to the tracking time! By performing the fix shown
above, we essentially scale it to the background time. This leads to a
~20 times larger value for the input signal hypothesis. Of course,
that means the optimization varying $g_{ae}$ will then have to scale
the parameter down even further resulting in a much lower limit (which
is then wrong of course).

This of course still means this commit is a major bug fix, because the
hardcoded factor of 10 is still rubbish.