[calibration] implement binned by time gas gain calculation
Handled by slicing the timestamps according to the `interval` field in
the config.toml file and fitting a polya for each slice. The values
are stored as attributes with indices attached.

Needs to be replaced by a dataset storing the gas gain values, because
overwriting (deleting) attributes is ridiculously slow.

Also needs to be cleaned up!