[plots] allow background rate plotting to cut to region
This adds the option to give input H5 files that were not created from
`likelihood.nim` using the gold region, but possibly the whole chip.

NOTE: this is dangerous in the sense that it still normalizes to the
gold region, always. So if a file with all data is given, but silver
region is selected, the normalization is wrong.
But of course it was always wrong in this case (except it just took
all data). So this doesn't make anything *worse*. Just makes it more
likely someone does something bad, because it now seems like the code
handles all regions correctly!