[logReader] change log file parsing, add VersionSchema
This is a huge change to the way we parse the data. The original code
was too slow when dealing with all of the historical CAST log files
for one thing.

Further, some fields either didn't make sense or were missing. Better
to store timestamps instead of durations from beginning of day.

Adds fields for log filenames and bad line counts.

The VersionSchema adds parsing of the `Version.idx` files for the slow
control logs to allow for (almost) sane parsing. Unfortunately, the
version schema promise is sometimes broken, either due to broken files
or because a file doesn't match its schema.. well.