[logL] allow to choose energy dset to use
This adds the option to allow the user to choose which energy dataset
should be used for the likelihood calculation.

It makes it possible to use the energyFromPixel dataset as the input
to the likelihood method, such that even for a detector for which a
ToT calibration is not available, the likelihood method can be used as
a rough approximation.