[Tpx3] improve parsing performance
Curiously, reusing the exact *same* sequences as buffers (i.e. not
using the intermediate `rawDataToDut` at all, makes
perfomance *plummet* for some reason...

In any case:
By far the limiting factor in performance here is the reading
performance from the H5 file. Apparently version 1.10 was very slow in
selecting hyperslabs and that was fixed in 1.12. However on 1.13 I get
nasty memory leaks, so uhh, for the time being we'll just ignore
higher performance (also means multithreading is not worth the
investment right now).