[karaPlot] add proper support for real custom plots
A custom plot is simply a plot that (either a histogram or scatter)
brings with it a `processData` procedure, which takes care of
- reading correct data
- generating the plot

The scatter plots buildable using the CLI are just a special case of
this.

Allows to quickly write a procedure that is called to generate a new plot.