[logL] improve line veto & septem veto logic
Under some circumstances a `passed` could be set back to `true` even
though it should not have:
- if a cluster *outside* the center chip *passes* the logL it would
set `passed` back to true! Despite the main cluster possibly *not*
passing!
Now we only check for "passed" if we have a cluster that is on the
center chip with its center.

Also add line veto information to the septem event display plots &
improves the logic to make it easier to reason about.