[logL] fix subtle bug septem pixel locations, use chip region lineveto
The pitch size used in combination with the hardcoded 14 mm size for
the Timepix lead to slight offsets for any chip with pixel numbers
larger than 255. That offset can lead to slight differences in things
like when the line veto is used etc.

Also we now actually use the input chip region for the line
veto. Note: the next commit will actually *fix* the line veto for
the full chip (broken as of this commit).