[Tpx3] do not fail if already read all data (last slice is empty)
The reason for this to happen is actually that if the last slice is
fully empty we can run into this as we will have read all data in the
dataset, but still have a `meta` slice element left. Therefore we will
try to read data, but `curIdx - oldIdx` will be 0, resulting in a
crash.