change dataset shapes in H5 from (N, 1) to (N, )
I was always under the impression that this is equivalent (and
required for vlen data). Therefore I always used is as such. But not
only is it not equivalent (or needed for vlen), if one *reads* using
an additional dimension with hyperslabs, reading becomes *orders of
magnitudes* slower!

This requires a new version of nimhdf5.