[utils] fix `chipGroups` iterator to never yield other run numbers
For the Timepix3 data I stumbled on a rather bad bug (that fortunately
shouldn't matter for Timepix1 data).

The old logic could identify a run such as "39" if the input group for
run "3" was given (because the prefix removal removed everything up to
the `9` and then the `scanf` for a run number `9` matched).

We now check whether it ends on a number (raw base path) or on a run
number (digit, full group of a run).