[karaPlot] improve logic detecting if a cut already applied
Before there was a chance of "dropping cuts" if a cut removed every
single entry in a run. Now we check if we "have cuts" in the first
place and otherwise only check for indices if we indeed have cuts. And
if no events left but we have cuts, we don't have to read any more data.