[karaPlot] make Cut/Region types objects, make them file / dset spec
The change from simple tuples to objects makes them more usable as
things that can be properly attached to a plot descriptor (via the
DataSelector type, as before).

We can add fields that map a cut explicitly to any number of
datasets (for which they should be applied) or for individual files,
so that when comparing files one can apply cuts specifically to only
some input files.