[raw] add ToT cut to remove pixels below & above specific ToT value
Especially for the current Tpx3 detector, ToT values at very low and
very high ranges appear in largish numbers that are unphysical. This
allows a convenient way to remove them.

The code also cleans up the Tpx1 parsing / logic for ToT removal that
was in place from the beginning somewhat.