[logL] simplify LikelihoodConfig obj, add CDL/55Fe matching via KS test
Simplifies many of the procedures by now taking a LikelihoodConfig
object instead of many (same) parameters.

Adds the logic described in our notes and developed in
`tools/determineEffectiveEfficiency.nim` to match 55Fe & CDL
distributions via an iterative approach using the KS test to check for
distribution matching.

There are still some TODOs, the implementation is not final with
respect to morphed inputs!