[fadc] rewrite FADC reconstruction code to be more flexible & correct
Fixes the old issue that the rotation of the spectra is wrong in
TPA. We were simply rotating the wrong direction. Needed `-nRoll`
instead of `nRoll`.

The rest of this deals with the fact that we now use Tensors for
freely in the FADC reconstruction & making sure the code can work with
seq & Tensor, but generally using Tensor directly. This avoids a bunch
of unnecessary copies that happened before.