[reco] fix usage of `fadc.high`, `len` in rise time using Tensor
High doesn't make sense here anymore! That was for the case of having
a dtype where high is the number of elements, in a 2D tensor it's not
well defined / refers to the "high" of the full flattened tensor.

The `len` has the same problem. Need `shape[0]` for a 2D tensor