[tools] make opt gas gain slice more generic, add more plots
This is a big change to the `optimizeGasGainSliceTime` tool for the
thesis. It first of all makes it more generic to work on any device
that has the required data files, an input path is now required.

Further, it adds many plots that make understanding the results
nicer, including discussions of goodness of fit tests on the median /
variance of the resulting distributions.