[logL] fix important bugs, refactor {.global.} by Config->Context
Fixes two important bugs that were introduced as a regression: A
`result` variable was shadowed by a local that wasn't seen in a
different scope and the name of the ingrid dataset kinds changed (or
rather their names are used nowadays).

The new LikelihoodContext now stores information that would otherwise have to
live inside of a {.global.}. Cleans up the code significantly.

There's still 3 globals in use in the code now. Need to be cleaned up too!