[raw] improve handling of list of raw files & name parsing
Instead of handing `seqs` manually we now have a `DataFiles` object
storing them and sorting them.

In addition the name parsing of the file names extracting event
numbers etc. is now improved to not get a string first and then parse
that, but rather we now use a custom parser for numbers after which a
`_` follows.