[cdl] add plots of all ingrid properties by CDL run
The plots compare the CDL data by run for the same target & filter
combination. Idea being that some tfKinds have very strong variations
in gas gain: so the question is whether that leads to a difference in
the InGrid properties? The answer seems to be a pretty certain no!
Fortunately, that means we will next replace the current logic by a
"by run" approach for the fits.