[logL] use fit based charge cuts from H5 file for logL in withCdlCuts
This implements the charge bounds as they are now written to the
`calibration-cdl*.h5` file to be used to cut to the *correct* clusters
that are actually present in the chosen CDL file.

Takes into account `fitByRun` as it was done in the CDL tool and uses
those charges for specific runs in that case.