[geometry] do not trust X is long Y is short cluster axis
In the real septemboard layout due to how we handle the data
apparently often the axes are switched (due to the mirroring of the
data I think). As we revert the mirroring it's not really of interest
to fix it, but it breaks the long and short axes definitions. It's
cleaner anyway to just use the actual *long* axis as the long one.

(Note that this likely explains why sometimes in the past we *did* see
some clusters that were mildly ε < 1!)