[FADC] improve calculation of baseline
Instead of calculating the FADC baseline based on median + 0.1 of the
max value we now calculate it from a truncated mean. We remove the
lower 30th percentile (in most cases more than the width of the
signal) and upper 5th.

This is still not perfect of course, but it gets the job done and
doesn't mean we need to implement something smarter. One could think
about using an approach that is effectively fitting local linear
functions and taking those parts that have a slope close to
zero. I.e. could be done using a Savitzky Golay filter with a
polynomial of order 1 and appropriate width. Then computing the
numerical gradient of the (much smoother) data and taking those
intervals of slope < some epsilon.