[FADC] do not compute rise/fall time to *baseline*, but slight offset
Instead of requiring to get back to the exact base line (which is
problematic due to individual noise in the bins) we now wait to reach
a slight offset of 2.5% below the baseline.

In the next commit this will also be combined with actually computing
things based on a moving average as well.