[logL] add `--readOnly` mode to likelihood for multiprocessing
The read only mode opens the input file (and CDL file) only in read
only mode. This means a couple of attributes are not written, but
generally these aren't very useful anyway in the input. Look into the
output for those of course.

Allows to implement multiprocessing on the likelihood files,
e.g. computing all the logL veto combinations in parallel.