[FADC] improve FADC rise/fall algo, change FADC data types
This is a big change.

First of all we change the data types that store the FADC data in
order to change the workflow in which the FADC data is
reconstructed. Instead of doing a weird mix of the reconstruction
where we perform the conversion from the raw FADC data to the
reconstructed, rotated, voltage based data together with the deduction
of the pulse minimum and noisy-ness, this is now split.

The main motivation for this change is that previously the FADC
reconstruction was tied to the noisy & minimum value calcs, which
meant that these were only performed when reconstructing from a _raw_
data file to a _reco_ data file. This made changing the algorithms of
the FADC parts very problematic, as there was no way to rerun only
e.g. the noise detection.

Secondly and more importantly, this changes the algorithm for the rise
and fall times even further. We now do what we alluded to in the past
and don't use a single fixed value as the starting point for the
rise/fall time calculations, but rather use the minimum of the
pulse, based on a mean of a percentile of the points around the
absolute minimum as a starting point. From this we then take the same
offset we need to "go down" from the baseline to as a starting point
from where to start counting rise / fall times. This essentially makes
sure that we only start counting the actual flank and not some shape
of the minimum of the pulse (which changes depending on pulse
height!).