[limit] add background interpolation code
This is straight from our notes in ~statusAndProgress~, i.e. it is
ugly and not cleaned up :)

The code performs the step to compute a k-d tree of the background
clusters to provide a smooth interpolation of the background rate at
any point on the chip. Plus a whole lot of stuff for "testing" etc.