[sim] massively improve fake event generation
We now have implemented generation of full fake events only based on
the gas gain and diffusion parameter of e.g. a run (or just sampled
from some range).

These reproduce the real data very well. The only important
differences are due to charge sharing effects, which we attempt to
simulate using neighbor pixel activation. Ideally that would not just
include the direct neighbors, but some exponential decay for further
pixels. The reason in reality is likely UV photons that hit the grid
and cause another avalanche.

It tries to use an H5 cache in `/dev/shm` for the diffusion values if
data is generated 'equivalent' to existing runs.

Some stuff is pretty hardcoded here still. That's a bit annoying but
it is what it is right now.