[NN] extend effective efficiency code drastically
It still contains _all sorts of_ experiments we did to try to
understand the behavior of our NN models. The most important stuff is
the code that's actually run as of this commit from the `main` proc.

Much (i.e. the slope and y intercept) is outdated experiments, but
should be committed at least for reference to our notes.