[limit] include NN signal eff, update default σ_s
This introduces the NN signal efficiency for the uncertainty &
software efficiency that goes into the amount of signal. Instead of
using the LnL software efficiency.

The `σ_s` value that is used by default for the limit program includes
the systematics for the LnL software efficiency. We took that out to
then compute the squared sum of the rest + the NN software efficiency.