[limit] update default `σ_sig` value from 4.2% to 2.7% (w/o ε_software)
This big change comes from using the _real distances_ of Sun ⇔ Earth
during the tracking time. The mean position will next be used to
increase the flux (real mean distance was about 98.9% of 1 AU) and the
systematic for it is from the standard deviation of all the distances
during tracking.