[limit] add support for limits as g⁴ or g_aγ² only
A `CouplingKind` is added to differentiate how we want to compute the
limit in terms of the coupling constant. The `ck_g_ae²·g_aγ²` case is
to compute the limit based on g_ae²·g_aγ² values directly instead of
fixing the g_aγ².

Note that this does not change the limit and hence the default remains
`g_aγ²`.

The case for `ck_g_aγ²` is in principle the major piece needed to
compute a pure axion-photon limit!