[plotting] rewrite background clusters plotting
We now use a table of pixel pairs and energies to store the
information from the files first and foremost. That can be turned into
the old count table if needed.

Now supports showing the energy of each cluster as an alternative and
even printing the text of the energy of each cluster (optionally only
in a radius around the center).