[karaPlot] adjust layouting of septemboard + FADC plots
Still a bit broken with the legend, but well.

The `|` is added because for some reason our left aligned text with
spaces does not take include the spaces anymore. I thought it was
either because of the ginger change that now uses the correct surface
type (e.g. PDF instead of PNG) for the text extent or the fact we
switched to width / height for text extents including using the
x/y_advance values, but neither seems to be it.