[ToT] fix parsing of ToT calibration files
The conversion we did before was based on what was done in the very
old TOS formats. But in our rewrite of TOCalib back in the day we
already fixed that. That was why the error bars were so tiny in the
ToT calibration plots.

(The conversion was essentially: 4 for 4 iterations, 256·256 for each
pixel of the chip)