[karaPlot] use `ggmulti` for InGrid+FADC, adjust spacing & getEnv
The spacing of the ingrid + fadc event display plots can now be adjust
with different environment variables! This shouldn't really be needed,
but was mainly added to debug the correct placement without
recompiling.

The ~USE_TEX~ env var can be used to activate plots to the TikZ
backend. ~SCALE~ can be used to scale everything up.