[tools] add mini tool to 'refresh' the run locals cache table
When reading and writing the *existing* file many times, it
accumulates a lot of "HDF5 junk". `h5stat` will show a lot of
unaccounted space.

We could probably do the same approach in the code itself, but for the
time being we just run it every once in a while.

The issue is that after multiple full applications of ~likelihood~
with a couple of different MLPs, the file can grow to >500MB in size
despite only containing a few kB of data.