[geometry] fix handling of rotation angle for septem board real layout
The inversion of the coordinates for non septemboard events means the
rotation angle needs to be inverted. We did not take care of *not*
inverting the rotation angle for septemboard events.

For the assignment of `rotationAngle` to the cluster geometry, we take
the absolute value. There is no more information than 0-180° degrees
anyway (as we don't try to predict the movement direction of tracks)